push

+++ install.sh
@@ -1,9 +1,11 @@
 #!/bin/bash -ex
 
-#
+# install openresty nodejs postgresql
 # Note: Need to run this as root
 #
-chmod 700 ~/.ssh/*
+
+# chmod 700 ~/.ssh/*
+
 install() {
   if [ -z $(which $1) ]; then
     sudo apt-get install -y $1
@@ -11,7 +13,7 @@ install() {
     echo "$1 already installed"
   fi
 }
-npm_install() {
+npm_global_install() {
   if [ -z $(which $1) ]; then
     sudo npm install -g $1
   else
@@ -19,102 +21,73 @@ npm_install() {
   fi
 }
 
-if [ ! -f /usr/local/openresty/bin/openresty ]; then
-  # add openresty latest
-  sudo apt-get -y install --no-install-recommends wget gnupg ca-certificates
-  wget -O - https://openresty.org/package/pubkey.gpg | sudo apt-key add -
-  echo "deb http://openresty.org/package/ubuntu $(lsb_release -sc) main" | sudo tee /etc/apt/sources.list.d/openresty.list
-else
-  echo "openresty already installed"
-fi
-
+# postgresql
 if [ -z "$(dpkg -l | grep -E '^ii\s+postgresql\s')" ]; then
   # add postgresql latest
   sudo sh -c 'echo "deb http://apt.postgresql.org/pub/repos/apt $(lsb_release -cs)-pgdg main" > /etc/apt/sources.list.d/pgdg.list'
   wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
-else
-  echo "postgresql already installed"
 fi
 
-sudo apt-get -q update
-
-install unzip
-install make
-install gcc
-
+# nodejs
 if [ -z $(which node) ]; then
   curl -fsSL https://deb.nodesource.com/setup_16.x | sudo -E bash -
-  sudo apt-get install -y nodejs
-else
-  echo 'nodejs 16 already installed'
 fi
 
-npm_install yarn
-npm_install env-cmd
-npm_install npm-check-updates
+# dotnet-sdk-6.0
+if [ -z $(which dotnet) ]; then
+  wget https://packages.microsoft.com/config/ubuntu/20.04/packages-microsoft-prod.deb -O packages-microsoft-prod.deb
+  sudo dpkg -i packages-microsoft-prod.deb
+  rm packages-microsoft-prod.deb
+fi
+
+# github cli
+if [ -z $(which gh) ]; then
+  curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg \
+  && sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg \
+  && echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null
+fi
 
 # install openresty
-OPENRESTY_PATH='/usr/local/openresty/bin:/usr/local/openresty/nginx/sbin:/usr/local/openresty/luajit/bin'
-export PATH=$OPENRESTY_PATH:$PATH
-BASH_RC=~/.bashrc
-if [ ! -f /usr/local/openresty/bin/openresty ]; then
-  sudo apt-get -y -q install openresty
-  service openresty stop
-  pgrep -x nginx && killall nginx
-  echo "openresty installed"
-  if [ ! -z $(grep $OPENRESTY_PATH $BASH_RC) ]; then
-    echo "openresty path already written to bashrc";
-  else
-    echo "export PATH=$OPENRESTY_PATH:\$PATH" >> ~/.bashrc;
-    echo "add PATH to ${BASH_RC}"
+PROJECT_DIR=$PWD
+OPENRESTY_VER=1.21.4.1
+OPENRESTY_DIR=/usr/local
+OPENRESTY_PATH="${OPENRESTY_DIR}/openresty/bin:${OPENRESTY_DIR}/openresty/nginx/sbin:${OPENRESTY_DIR}/openresty/luajit/bin"
+PATH="$OPENRESTY_PATH:$PATH"
+BASH_RC=~/.profile
+if [ ! -z "$(grep openresty/bin $BASH_RC)" ]; then
+  echo "openresty path already written to ${BASH_RC}";
+else
+  echo "export PATH=$OPENRESTY_PATH:\$PATH" >> ${BASH_RC};
+  echo "add $OPENRESTY_PATH to ${BASH_RC}"
+fi
+if [ ! -f ${OPENRESTY_DIR}/openresty/bin/openresty ]; then
+  cd /tmp
+  if [ ! -f openresty-${OPENRESTY_VER}.tar.gz ]; then
+    wget https://openresty.org/download/openresty-${OPENRESTY_VER}.tar.gz
   fi
-  # mkdir lua_modules
+  if [ ! -d openresty-${OPENRESTY_VER}/ ]; then
+    tar -xvf openresty-${OPENRESTY_VER}.tar.gz
+  fi
+  cd openresty-${OPENRESTY_VER}/
+  #  --with-http_geoip_module
+  ./configure --prefix=${OPENRESTY_DIR}/openresty --with-pcre-jit -j8
+  make -j4 && make install
+  # service openresty stop
+  # pgrep -x nginx && killall nginx
   opm get ledgetech/lua-resty-http
   opm get bungle/lua-resty-template
   opm get fffonion/lua-resty-openssl
   opm get xiangnanscu/pgmoon
-  opm get xiangnanscu/lua-resty-rax
-  opm get xiangnanscu/lua-resty-array
-  opm get xiangnanscu/lua-resty-object
   opm get xiangnanscu/lua-resty-inspect
 else
   echo "openresty already installed"
 fi
 
-
-
-if [ -z "$(dpkg -l | grep -E '^ii\s+postgresql\s')" ]; then
-  [ -z $PG_PASSWORD ] && echo "Plase set PG_PASSWORD" && exit 1
-  # install postgresql
-  sudo apt-get -y install postgresql
-  service postgresql start
-
-  PG_VERSION_VERBOSE=`pg_config --version`
-  PG_VERSION=${PG_VERSION_VERBOSE:11:2}
-
-  echo "pg version: $PG_VERSION"
-
-  # modify setting
-  sed -i 's/max_connections = 100/max_connections = 400/g' /etc/postgresql/$PG_VERSION/main/postgresql.conf
-  sed -i 's/shared_buffers = 128MB/shared_buffers = 256MB/g' /etc/postgresql/$PG_VERSION/main/postgresql.conf
-  sed -i 's/#password_encryption = scram-sha-256/password_encryption = md5/g' /etc/postgresql/$PG_VERSION/main/postgresql.conf
-  # sed -i 's/"scram-sha-256"/@@@/g' /etc/postgresql/$PG_VERSION/main/pg_hba.conf
-  sed -i 's/\(\s\)scram-sha-256/\1md5/g' /etc/postgresql/$PG_VERSION/main/pg_hba.conf
-  # sed -i 's/@@@/"scram-sha-256"/g' /etc/postgresql/$PG_VERSION/main/pg_hba.conf
-  echo "modify postgresql settings done"
-  service postgresql restart
-  sudo -u postgres psql -w postgres <<EOF
-    ALTER USER postgres PASSWORD '$PG_PASSWORD';
-EOF
-
-else
-  echo "postgresql already installed"
-fi
-
-LUAROCKS_VER='3.8.0'
+# luarocks
+LUAROCKS_VER='3.9.1'
 LUAROCKS_FD=luarocks-${LUAROCKS_VER}
 LUAROCKS_GZ=${LUAROCKS_FD}.tar.gz
-if [ ! -f /usr/local/openresty/luajit/bin/luarocks ]; then
+if [ ! -f ${OPENRESTY_DIR}/openresty/luajit/bin/luarocks ]; then
   cd /tmp
   if [ ! -f $LUAROCKS_GZ ]; then
     wget https://luarocks.org/releases/$LUAROCKS_GZ
@@ -123,13 +96,15 @@ if [ ! -f /usr/local/openresty/luajit/bin/luarocks ]; then
     tar zxpf $LUAROCKS_GZ
   fi
   cd $LUAROCKS_FD
-  ./configure --prefix=/usr/local/openresty/luajit \
-      --with-lua=/usr/local/openresty/luajit/ \
+  ./configure --prefix=${OPENRESTY_DIR}/openresty/luajit \
+      --with-lua=${OPENRESTY_DIR}/openresty/luajit/ \
       --lua-suffix=jit \
-      --with-lua-include=/usr/local/openresty/luajit/include/luajit-2.1
+      --with-lua-include=${OPENRESTY_DIR}/openresty/luajit/include/luajit-2.1
   make && make install
   echo "luarocks ${LUAROCKS_VER} installed"
+
   # luarocks install  --tree lua_modules luasocket
+  cd $PROJECT_DIR
   luarocks install luasocket
   luarocks install luafilesystem
   luarocks install luacheck
@@ -140,20 +115,62 @@ if [ ! -f /usr/local/openresty/luajit/bin/luarocks ]; then
 else
   echo "luarocks already installed"
 fi
+
+sudo apt-get -q update
+
+install nodejs
+install unzip
+install make
+install gcc
+
+npm_global_install yarn
+npm_global_install env-cmd
+
+if [ -z "$(dpkg -l | grep -E '^ii\s+postgresql\s')" ]; then
+  if [ -z $PG_PASSWORD ]; then
+    read -p "please provide postgresql password for user postgres:" PG_PASSWORD
+  fi
+  # install postgresql
+  sudo apt-get -y install postgresql
+
+  PG_VERSION_VERBOSE=`pg_config --version`
+  PG_VERSION=${PG_VERSION_VERBOSE:11:2}
+
+  echo "pg version: $PG_VERSION"
+
+  # modify setting
+  sudo sed -i 's/max_connections = 100/max_connections = 400/g' /etc/postgresql/$PG_VERSION/main/postgresql.conf
+  sudo sed -i 's/shared_buffers = 128MB/shared_buffers = 256MB/g' /etc/postgresql/$PG_VERSION/main/postgresql.conf
+  sudo sed -i 's/#password_encryption = scram-sha-256/password_encryption = md5/g' /etc/postgresql/$PG_VERSION/main/postgresql.conf
+  sudo sed -i 's/\(\s\)scram-sha-256/\1md5/g' /etc/postgresql/$PG_VERSION/main/pg_hba.conf
+
+  sudo service postgresql restart
+  sudo -u postgres psql -w postgres <<EOF
+    ALTER USER postgres PASSWORD '$PG_PASSWORD';
+EOF
+
+else
+  echo "postgresql already installed"
+fi
+
+# install dotnet-sdk-6.0
+if [ -z $(which dotnet) ]; then
+  wget https://packages.microsoft.com/config/ubuntu/20.04/packages-microsoft-prod.deb -O packages-microsoft-prod.deb
+  sudo dpkg -i packages-microsoft-prod.deb
+  rm packages-microsoft-prod.deb
+  sudo apt-get update && sudo apt-get install -y dotnet-sdk-6.0
+fi
+
+# install github cli
+if [ -z $(which gh) ]; then
+  curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg \
+  && sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg \
+  && echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null \
+  && sudo apt update \
+  && sudo apt install gh -y
+fi
+
 git config --global user.email "280145668@qq.com"
 git config --global user.name "xiangnan"
 git config --global receive.denyCurrentBranch updateInstead
 git config --global receive.advertisePushOptions true
-
-# install dotnet-sdk-6.0
-wget https://packages.microsoft.com/config/ubuntu/20.04/packages-microsoft-prod.deb -O packages-microsoft-prod.deb
-sudo dpkg -i packages-microsoft-prod.deb
-rm packages-microsoft-prod.deb
-sudo apt-get update && sudo apt-get install -y dotnet-sdk-6.0
-# install github cli
-curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg \
-&& sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg \
-&& echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null \
-&& sudo apt update \
-&& sudo apt install gh -y
-
